kafka and zookeeper added, advanced log system created between services and some improvements

+++ CompanyService/src/main/java/com/merensekkeli/companyservice/controller/contract/impl/CompanyControllerContractImpl.java
@@ -3,15 +3,19 @@ package com.merensekkeli.companyservice.controller.contract.impl;
 import com.merensekkeli.companyservice.controller.contract.CompanyControllerContract;
 import com.merensekkeli.companyservice.dto.CompanyDTO;
 import com.merensekkeli.companyservice.entity.Company;
+import com.merensekkeli.companyservice.exception.ItemNotFoundException;
+import com.merensekkeli.companyservice.general.KafkaProducerService;
 import com.merensekkeli.companyservice.mapper.CompanyMapper;
 import com.merensekkeli.companyservice.request.CompanySaveRequest;
 import com.merensekkeli.companyservice.service.CompanyEntityService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Slf4j
@@ -19,12 +23,16 @@ import java.util.List;
 public class CompanyControllerContractImpl implements CompanyControllerContract {
 
     private final CompanyEntityService companyEntityService;
+    private final KafkaProducerService kafkaProducerService;
+    @Value("${application.name}")
+    private String appName;
 
     @Override
     public CompanyDTO saveCompany(CompanySaveRequest request) {
         Company company = CompanyMapper.INSTANCE.convertToCompany(request);
         company = companyEntityService.save(company);
         log.info("Company saved with id: {}", company.getId());
+        kafkaProducerService.sendMessage("infoLog", appName, "Company saved with id: " + company.getId());
         return CompanyMapper.INSTANCE.convertToCompanyDTO(company);
     }
 
@@ -51,19 +59,28 @@ public class CompanyControllerContractImpl implements CompanyControllerContract 
 
         company = companyEntityService.save(company);
         log.info("Company updated with id: {}", company.getId());
+        kafkaProducerService.sendMessage("infoLog", appName, "Company updated with id: " + company.getId());
         return CompanyMapper.INSTANCE.convertToCompanyDTO(company);
     }
 
     @Override
     public void deleteCompany(String id) {
+        Optional<Company> company = companyEntityService.findById(id);
+
+        if (company.isEmpty()) {
+            log.error("Company not found with id: {}", id);
+            throw new ItemNotFoundException("Company not found with id: " + id);
+        }
         companyEntityService.delete(id);
         log.info("Company deleted with id: {}", id);
+        kafkaProducerService.sendMessage("infoLog", appName, "Company deleted with id: " + id);
     }
 
     @Override
     public List<CompanyDTO> findByNamed(String searchItem) {
         List<Company> companyList = companyEntityService.findByNamed(searchItem);
         log.info("Companies found with search item: {}", searchItem);
+        kafkaProducerService.sendMessage("infoLog", appName, "Companies found with search item: " + searchItem);
         return companyList.stream().map(CompanyMapper.INSTANCE::convertToCompanyDTO).toList();
     }
 
@@ -71,13 +88,26 @@ public class CompanyControllerContractImpl implements CompanyControllerContract 
     public List<CompanyDTO> findByProperty(String searchItem) {
         List<Company> company = companyEntityService.findByProperties(searchItem);
         log.info("Company found with search item: {}", searchItem);
+        kafkaProducerService.sendMessage("infoLog", appName, "Company found with search item: " + searchItem);
         return company.stream().map(CompanyMapper.INSTANCE::convertToCompanyDTO).toList();
     }
 
     @Override
     public CompanyDTO getCompanyById(String id) {
         Company company = companyEntityService.findByIdWithControl(id);
-        log.info("Company found with id: {}", id);
+        return CompanyMapper.INSTANCE.convertToCompanyDTO(company);
+    }
+
+    @Override
+    public CompanyDTO getCompanyByIdOuther(String id) {
+        Company company;
+        try {
+            company = companyEntityService.findByIdWithControl(id);
+        } catch (ItemNotFoundException e) {
+            log.error("Company not found with id: {}", id);
+            return null;
+        }
+
         return CompanyMapper.INSTANCE.convertToCompanyDTO(company);
     }
